requirements/common.txt
Django==1.10.5
psycopg2==2.7.3
pillow==4.0.0
django-simple-history==1.8.2
adal==0.4.5
djangorestframework==3.6.3
requests==2.18.2
